+++ core/src/no/ntnu/game/Models/AbstractPowerUpFactory.java
@@ -1,5 +1,7 @@
 package no.ntnu.game.Models;
 
+import no.ntnu.game.Factory.PowerUp.PowerUp;
+
 public interface AbstractPowerUpFactory {
     PowerUp createPowerUp();
 }
