+++ core/src/no/ntnu/game/Models/TreePart.java
@@ -1,11 +1,14 @@
 package no.ntnu.game.Models;
 
 import com.badlogic.gdx.graphics.Color;
+
+import no.ntnu.game.Factory.PowerUp.PowerUp;
+
 public class TreePart {
-            String value;
-            Color color;
-            float x,y;
-            PowerUp powerup;
+    String value;
+    Color color;
+    float x, y;
+    PowerUp powerup;
 
     public String getValue() {
         return value;
@@ -27,22 +30,22 @@ public class TreePart {
         return powerup;
     }
 
-    public TreePart(String value , Color color){
+    public TreePart(String value, Color color) {
 
-            this.value = value;
-            this.color = color;
-            this.x = 100;
-            this.y = 100;
-            this.powerup = null;
+        this.value = value;
+        this.color = color;
+        this.x = 100;
+        this.y = 100;
+        this.powerup = null;
 
-        }
-
-        public void setPos(float x , float y){
-                this.x =x;
-                this.y =y ;
     }
 
-        public void setPowerup(PowerUp powerup){
-                this.powerup = powerup;
-        }
+    public void setPos(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public void setPowerup(PowerUp powerup) {
+        this.powerup = powerup;
+    }
 }
